+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/SdkHarnessRegistry.java
@@ -21,7 +21,7 @@ import org.apache.beam.model.pipeline.v1.Endpoints.ApiServiceDescriptor;
 import org.apache.beam.runners.fnexecution.control.FnApiControlClient;
 import org.apache.beam.runners.fnexecution.data.GrpcDataService;
 import org.apache.beam.runners.fnexecution.state.GrpcStateService;
-import org.apache.beam.sdk.fn.GrpcFnServer;
+import org.apache.beam.sdk.fn.server.GrpcFnServer;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 /** Registry used to manage all the connections (Control, Data, State) from SdkHarness */
